refactor(cli-test): split DatashareCommandTest by CLI section
